requirements.txt
python-dotenv
pytesseract
pdfplumber
PyMuPDF
faiss-cpu
sentence-transformers
langchain
numpy
